refactor subtitle-annotation.ts into multiple files
