--- common/subtitle-annotations/index.ts
@@ -1,1 +1,0 @@
-export * from './subtitle-annotations';
